Fix for Issue #97 - Data18 Web Content not scraping posters in europe
The pulled in images were using a different IP than the american version
of the page and were giving response code 302 instead of 200 when
downloading them. This prevented the images from actually downloading
and so the scraper didn't add them to the nfo. I've switched over to
just using the American IP address with a manual subsitution of the IP
address, for now.