fix writing metadata when source is a file
we used `newMovieFile` as the destination path for the metadata files
whether the movie was a file, or a directory.

If it is a directory, that is okay, the directory is renamed, then
metadata is written to it.

If it is a file though, that breaks, because the content of
`newMovieFile` after rename is a file path, not a directory path.

Fix: use `newMovieDirectoryFile` as the destination path instead
and define it in both branches, whether the source is a file or a directory.